feat: extend schema to include other items
