+++ crates/cheatcodes/src/defs/vm.rs
@@ -2,7 +2,7 @@
 // module. Instead, we emit custom diagnostics in `#[derive(Cheatcode)]`.
 #![allow(missing_docs)]
 
-use crate::{Cheatcode, CheatcodeDef, Group, Mutability, Safety, Status, Visibility};
+use super::*;
 use alloy_sol_types::sol;
 use foundry_macros::Cheatcode;
 
@@ -20,7 +20,7 @@ sol! {
 interface Vm {
     //  ======== Types ========
 
-    /// Error thrown by a cheatcode.
+    /// Error thrown by cheatcodes.
     error CheatcodeError(string message);
 
     /// A modification applied to either `msg.sender` or `tx.origin`. Returned by `readCallers`.
